# Improve performance of labels comparison when querying store-gateways