#### What this PR does

This PR applies https://github.com/grafana/mimir/pull/5836 to two additional places that could benefit from it.

This improves performance when using `stringlabels` (as Mimir does). Note that the existing benchmark includes iterating through all of the samples in the series, which is why there isn't a dramatic latency change in the results below:

```
goos: darwin
goarch: arm64
pkg: github.com/grafana/mimir/pkg/querier
                                    │ before.txt  │             after.txt             │
                                    │   sec/op    │   sec/op     vs base              │
_blockQuerierSeriesSet_iteration-10   812.2m ± 0%   814.8m ± 0%  +0.32% (p=0.009 n=6)

                                    │  before.txt  │             after.txt              │
                                    │     B/op     │     B/op      vs base              │
_blockQuerierSeriesSet_iteration-10   8.426Mi ± 0%   7.694Mi ± 0%  -8.69% (p=0.002 n=6)

                                    │ before.txt  │             after.txt             │
                                    │  allocs/op  │  allocs/op   vs base              │
_blockQuerierSeriesSet_iteration-10   256.0k ± 0%   240.0k ± 0%  -6.25% (p=0.002 n=6)
```



#### Which issue(s) this PR fixes or relates to

(none)

#### Checklist

- [n/a] Tests updated
- [n/a] Documentation added
- [x] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`